fix(er): chained exception frame ordering
We fix the order in which frames are collected from an exception chain.
We make sure that the frames closer to the point where the initial
exception is thrown are captured first. This ensures that we propagate
the tags on the relevant spans as the exception chain bubbles up and
unwinds.