fix(llmobs): treat google adk tools as either `BaseTool` or generic function (#15402)
## Description

Previously, for the Google ADK LLM Observability integration, we would
try and extract tool name and descriptions as if they were all
`BaseTool`s, when really they can also be a generic function. This PR
incorporates that logic.

Fixes https://github.com/DataDog/dd-trace-py/issues/15174

## Testing

Changing one of the tools in the tests to a function instead of a
`FunctionTool` reproduced this issue in tests (error + missing span
event).

## Risks

None 

## Additional Notes

We should in the future update the Google ADK integration as I believe
we currently only capture the top-level runner span as the agent when we
should also capture sub agents.