chore(ci_visibility): reabsorb ddtestpy into ddtrace (#15396)
## Description

This PR moves all the code from https://github.com/DataDog/dd-test-py
back into ddtrace. The goal is to eventually split it off again, after
we have the right abstractions inside ddtrace to make it happen.

- `ddtestpy/*` was moved into `ddtrace/testing/`.
- `tests/*` was moved into `tests/testing/` (I know, not the greatest
name. Suggestions accepted. I have considered `test_optimization`, but I
expect more things to get there that are related to testing but not
necessarily Test Optimization, such as code coverage. Also I'm trying to
avoid using product names as the main module name, given the past
history of product renames.)
- The ddtrace pytest entry point was changed to a new module
`ddtrace.testing.internal.pytest.entry_point`, which imports either the
old plugin or the new one depending on the value of the
`DD_PYTEST_USE_NEW_PLUGIN` variable. It defaults to false for now, but
the idea is to switch it to default to true once the new plugin is
feature-complete and more battle-tested. This is similar to the way we
introduced the existing (v2) pytest plugin back in ddtrace 2.x.
- Imports were changed according to the new code location.
- The `--ddtestpy` and similar pytest command line options were renamed
to `--ddtrace`. `--ddtestpy-with-ddtrace` was renamed to
`--ddtrace-patch-all`.
- The vendored ddtrace code coverage modules that we had in ddtestpy
were removed; we just use the modules from ddtrace directly.
- The `ci` test fixtures were removed, as they are exactly the same as
the ones in `tests/tracer/fixtures/ci`.
- Likewise, the `ddtestpy.internal.codeowners` module was removed as it
was virtually identical to `ddtrace.internal.codeowners` (the only
differences were mypy-related).
- Code that depends on the rest of ddtrace internals was moved to
`ddtrace.testing.internal.tracer_api`, to make it easier to keep track
of what are the dependencies between the testing code and everything
else.
- Riot environments were added for the new code.
- Suitespec was updated accordingly.

## Testing

Unit tests were brought together with the code. We will add more in
subsequent PRs.

Manual testing on a local  repo.

## Risks

None, as the new plugin is not enabled by default.

## Additional Notes

I'm not adding a release note in this PR. We will do so once the plugin
is ready for wider usage.

---------

Co-authored-by: Brett Langdon <brett.langdon@datadoghq.com>