chore(langchain): make langchain runnable lambda test less flaky (#15512)
## Description

One of the tests added in #15477 asserts a list of span links, but since
we only care about content and not order, asserting the raw lists
sometimes flaked if the list was out of order (which doesn't matter for
span links). so instead, since we cannot hash them into a set because
they're dicts, we assert one expected element at a time.

## Testing

updating the test itself, will check on the health of it after it merges

## Risks

none, passed still locally