chore(profiling): improve typing in _asyncio.py (#15506)
## Description

https://datadoghq.atlassian.net/browse/PROF-13197

This improves typing for the `_asyncio.py` module. Note that the result
is somewhat verbose, which I'm not a fan of, but honestly I'm not quite
sure how to make things much better.

Two ideas I had for `typing`:
* `import typing as t` to make lines shorter... 
* ... or `from typing import Any, Callable, ...` to get rid of the
qualification altogether

For `asyncio`:
* Currently I `import asyncio as aio` for type checking, which isn't
great since we _also_ import it as `asyncio` (but can't use the
`asyncio` name in the import wrapper...)
* We could do the same and do `from asyncio import Task, ...` for type
checking. I think this could be OK.