chore(llmobs): [MLOB-4814] remove tagging pydantic dependencies on agent manifest (#15490)
## Description

We received a customer issue about high memory use while using our
Pydantic AI integration. After investigating the issue, it seems that
this was caused by trying to serialize a large dependencies object to
tag onto the span.

This PR avoids tagging agent dependencies on Pydantic AI agent spans.
This prevents potential memory issues caused by the serialization that
is required to tag these objects onto spans. We are currently not
exposing these dependencies anywhere in the UI, so this should be a
benign change.

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->