refactor(profiling): change how `GenInfo::is_running` works (#15508)
## Description

This change replicates https://github.com/P403n1x87/echion/pull/202.

This PR changes the way `GenInfo::is_running` works. Previously, it
would indicate whether the _current_ coroutine was on CPU; now, it
indicates whether the current coroutine _or the coroutine it
(recursively) awaits_ is on CPU.
Making that change also allows us to do less work when we check whether
the current coroutine is on CPU or not. Because a Coroutine / Generator
/ `GenInfo` can only be running if it is not awaiting another Generator,
we do not need to compute `is_running` if we have `await != nullptr`
(and we take `await->is_running` for the value in that case).

Making that change makes it easier to check whether a Task is currently
on-CPU; and allows to do less work when we decide how to unwind
`asyncio` Tasks (cf the changes in `TaskInfo` which doesn't need the
`is_on_cpu` method that iterates on the `await` chain anymore).

Note that I checked whether `GenInfo::is_running` was used in any other
way than the one I describe and simplify; it is not, so I do think this
change is safe to make as-is.

**Note** This PR makes sense on its own, but it is in the context of
making https://github.com/P403n1x87/echion/pull/198 simpler.

**Note** This doesn't need a changelog entry as it makes no difference
to the user, it's purely a refactor.