feat(profiling): add `process_id` tag (#15515)
## Description

This PR adds a new `process_id` tag to Profiles captured by dd-trace-py.
The value of this tag is the PID of the current process. Due to the way
it is populated (in the `UploaderBuilder`, and not once per process), it
is by definition updated automatically in forks.

Note that we could also have a `set_process_id` function in `ddup` (as
opposed to the function determining the PID itself), but I figured the
native implementation would be more efficient than having to marshall a
Python string object.