fix(ray): increase ray instrumentation stability (#15446)
Ray integration was until very soon only tested internally. We had
several issues with instrumenting jobs based on when ray was initialized
and some crashes related to context injection.

This PR is created to fix all the small issues and make the integration
stable. Ray integration should be advertised to be used only on versions
including this fix. The releasenote is covering the only customer facing
issue.

## Context 

During internal dogfooding some bugs were found/some assumptions were
challenged:
- Most of the users can't called `ray.init()` at the top of the file
(which was a prerequisite for a fully working instrumentation).
After analyzing all the consequences of calling ray.init() late or
waiting for auto-initialization (for example when a task is submitted
for the first time), it appears that the only viable solution was to
change the way we instrument the cluster.
- Some ray internal tasks were instrumented and created non
actionnable/useful orphan spans.
- We still had issues with context injection
- Service names were not properly working when unspecified.

## Changes

- Update docs to recommend using ddtrace-run. Note that we advise to
deactivate some integrations otherwise instrumenting ray cluster is very
noisy. We will be able to remove that if we make an upstream
contribution.
- Fix service names and specify precedence order in docs. 
- Prevent some potential future bugs by not instrumenting cross_language
tasks
- Add a fix for injecting `dd_trace_ctx` in tasks. 

## Testing

I added some tests to ensure all the behavior are tested:
- Added some tests related to service name
- Added some tests to show the issues with not calling ray.init() to be
sure it is documented and any changes will be detected. This tests are
also ensuring that nothing crashes.

## Additional Notes

Ray test suite were run while ddtrace activated increasing the
confidence in this PR.