chore(ci_visibility): set span type to "test" for ddtrace integrations (#15526)
## Description

The Selenium integration checks if the root span is of `test` type
before adding its tags. This PR adjusts the span type of the ddtrace
span created by the new pytest plugin.

Additionally, fix the `selenium` riot environment to use the correct env
var to set the test agent port.

## Testing

Manual testing to see that the Selenium integration works with the new
plugin.

Currently we can't do snapshot tests with the new plugin because the
test agent does not support citestcycle traces sent via EVP proxy mode,
and the new plugin does not support the legacy mode.

## Risks

None.

## Additional Notes

None.