refactor(profiling): make `StringTable::lookup` `const` (#15547)
## What does this PR do?

This PR updates the `StringTable::lookup` method to be `const`. This
required making the `std::mutex` we have `mutable`, which is OK –
`mutable` is typically used for that.

Additionally, it also makes the return type
`std::reference_wrapper<const std::string>` as I don't think we should
allow callers of `lookup` to change the underlying data, which the
previous signature would allow.

>
[[cppreference]](https://en.cppreference.com/w/cpp/language/cv.html#mutable)
mutable is used to specify that the member does not affect the
externally visible state of the class (as often used for mutexes, memo
caches, lazy evaluation, and access instrumentation).

Echion PR: https://github.com/P403n1x87/echion/pull/209