chore(ci_visibility): add pytest-benchmark support to new plugin (#15541)
## Description

Add support for the pytest-benchmark plugin to the new pytest plugin.
When the `benchmark` fixture from
[pytest-benchmark](https://pypi.org/project/pytest-benchmark/) is in
use, we extract statistics generated by the plugin and add them as
tags/metric to the test run event.

This is based on
https://github.com/DataDog/dd-trace-py/blob/v4.0.0/ddtrace/contrib/internal/pytest/_benchmark_utils.py,
but with a lot less indirection, and an explicit separation between
(string) tags and (float) metrics.

## Testing

Unit tests.

## Risks

None.

## Additional Notes

None.