chore(ci_visibility): add pytest-bdd support to new plugin (#15542)
## Description

Add support for the pytest-bdd plugin to the new pytest plugin.

This code is basically transplanted from
https://github.com/DataDog/dd-trace-py/blob/v4.0.0/ddtrace/contrib/internal/pytest/_pytest_bdd_subplugin.py,
adapted to work with the new plugin.

The tests are likewise adapted from
https://github.com/DataDog/dd-trace-py/blob/v4.0.0/tests/contrib/pytest_bdd/test_pytest_bdd.py.

The logic for adding hooks conditionally depending on which plugins are
available was changed. Previously, the only such case we had was xdist,
and we did it by subclassing `TestOptPlugin` with the added xdist hooks.
This is not feasible anymore now that we have multiple plugins with
added hooks (xdist and pytest-bdd). Instead, the logic was changed so
that the subplugins are instantiated separately after the main plugin,
and receive a reference to the main plugin when initialized. This allows
any set of subplugins to be activated independently.

## Testing

Unit tests were copied over from the old plugin.

## Risks

None.

## Additional Notes

One known issue with the new implementation is that it requires pytest
to be called with `--ddtrace-patch-all` for the pytest-bdd spans to be
collected properly (because the new plugin's span processor only gets
installed if `--ddtrace-patch-all` is used). This will be addressed in a
separate PR.