chore(tracer): prevent some fatal failures on metastruct serialisation (#15551)
## Description

If any instrumentation put some non serializable object in the
metastruct of a span, currently, the span is never sent due to a
serialization error. This may be hard to understand and debug.
This PR replaces the most common serialization errors by an error string
that will be sent in the serialized data.

## Testing

metastruct test was updated to include non serializable object and
overflow integer

APPSEC-60180