ci: add precomputed pipeline variables (#15582)
## Description

This PRs adds an initial "setup" stage to the pipeline where we compute
and expose the following variables to the rest of the pipeline:


- **GH_PR_NUMBER:** The GitHub PR number if there is an open PR for this
commit, empty otherwise
- **HAS_OPEN_PR:** "true" if there is an open PR for this commit,
"false" otherwise
- **IS_MAIN_BRANCH:** "true" if the current branch is main
- **IS_RELEASE_BRANCH:** "true" if the current branch is a release
branch (e.g., "1.2"), "false" otherwise
- **IS_RELEASE:** "true" if the current commit is a release tag (e.g.,
"v1.2.3"), "false" otherwise
- **IS_MERGE_QUEUE:** "true" if the current branch is a merge queue
branch (e.g., starts with "gh-readonly-queue/"), "false" otherwise

The goal is to provide consistent controls for which jobs to run when,
e.g. should a job run only when we have a PR open or only during a
release?

real example, with #15572 we want to be sure we still restrict
benchmarks to only run/pass when we have an open PR to reduce volume,
today that is controlled by the fact that the GHA wheel building only
happens when a PR is open.

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->