feat(crashtracking): support runtime stack collection for all Python versions (#15568)
## Description

This PR allows for runtime stacks collection for all CPython versions
greater than 3.9.*

We are okay with using dangerous Python APIs in a crash context. This is
because the Crashtracker has safety nets that ensure customer experience
will not be impacted; at worst, only the `runtime_stacks` section, which
is experimental, will be missing from a crash report.

We don't support 3.9 because this will require special linking logic to
CPython symbol as `PyThreadState_GetFrame` is not in the public API
until 3.10 and `dd-trace-py` will drop support for 3.9 soon anyways.

## Testing

Uploaded some crashes to telemetry logs. Runtime stacks unit test has
also been updated to check for `string_at` in the experimental section
of the crash report for all versions > 3.9


## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->