chore(ci_visibility): add xfail/xpass support to the new plugin (#15601)
## Description

Add support for xfail/xpass to the new plugin, with the `test.report`
tag set to `xfail` or `xpass`, and `pytest.xfail.reason` set to the
xfail reason. Also propagate the `wasxfail` status from retries to the
final report.

The behavior is not identical to the old plugin. xfailed failing tests
are reported as skipped by pytest; xfailed passing tests as passed when
not in xfail_strict mode, and as failed when in xfail_strict mode. The
old plugin has some [quite convoluted
logic](https://github.com/DataDog/dd-trace-py/blob/v4.0.0/ddtrace/contrib/internal/pytest/_plugin_v2.py#L757)
that tries to second-guess the "actual" pass/fail/skip status from an
xfailed test (so for example a skipped xfail test is reported to the
backend as passed). The new plugin just uses the status reported by
pytest as the test status, so if pytest reports the test as skipped, we
report it as skipped, and so on. The `test.result` tag will still
contain the `xfail`/`xpass` value as usual.

## Testing

Unit tests added. Some existing unit tests were refactored to use an
actual TestReport instead of a mock.

## Risks

None.

## Additional Notes

None.