fix: ensure connection is reset if from_http_response raises an exception (#15559)
## Description

Previously, the http writer was calling `from_http_response` in the
`else` block of the exception handler that does the request, set's the
log level, et cetera. However this could have the problem of
`from_http_response` itself raising an exception (e.g.
`http.client.IncompleteRead` as seen on dogweb staging recently) outside
of the `Exception` handler, which would leave the `HTTPWriter` object in
need of a call to `_reset`; the `finally` would call `_reset` but only
if `reuse_connection` aka `DD_TRACE_WRITER_REUSE_CONNECTIONS` was
enabled, which is not the default.

## Testing
New tests have been added.

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->

---------

Signed-off-by: Juanjo Alvarez <juanjo.alvarezmartinez@datadoghq.com>
Co-authored-by: Quinna Halim <quinna.halim@datadoghq.com>