fix(openai): remove openai API key parsing entirely (#15626)
## Description
Closes #15585.

This PR removes API key parsing from the openai integration for two
reasons:
1. We were not using this anywhere (we used to tag the last 4 digits of
the API key similar to how OpenAI formats theirs in UI, but tagging this
was removed quite some time ago)
2. This caused errors in use cases involving `azure-identity` that
involved API keys as functions (i.e. not directly storing API keys in
app/memory), because we always assumed this was a string.

<!-- Provide an overview of the change and motivation for the change -->
## Testing
<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->