refactor(profiling): remove unused return value (#15521)
## Description

`pyframe_to_frames()` returns the number of frames that are present in
the original traceback, but it is no longer used.

It seems like it was there to handle the case where we truncated `n`
number of frames when we were using Python exporter. Given that we
switched over to libdatadog exporter, which handles frame truncation in
a slightly different manner, I'll follow up with another PR for that
fix.
 

<!-- Provide an overview of the change and motivation for the change -->

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->