chore(ci_visibility): always enable trace filter in new plugin (#15631)
## Description

Previously we only enabled the trace filter (which captures non-test
spans and translates them into test events) when `--ddtrace-patch-all`
was used. The idea was that only when ddtrace integrations are used we
would have ddtrace spans.
But we actually need the filter to always be on to ensure that anything
that generates spans during a test gets captured (and not sent to the
regular APM writer). Moreover, pytest-bdd generates ddtrace spans even
if no other ddtrace integration is in use. This PR makes it always on,
and makes enabling the trace filter an independent operation from
enabling ddtrace integrations (which is what `--ddtrace-patch-all`
does).

## Testing

`--ddtrace-patch-all` was removed from pytest-bdd tests that used to
require it, and they still pass (they wouldn't before this change).

## Risks

None.

## Additional Notes

None.