chore(ci_visibility): track EFD abort reason in new plugin (#15641)
## Description

Set the `test.early_flake.abort_reason` to `slow` if a test that was
supposed to be retried by EFD takes longer than 5 minutes to run.

Also set `test.early_flake.abort_reason` on the test session to `faulty`
if the EFD session is faulty.

This PR also fixes a number of logic errors:
- The initial attempt was being sent twice to the writer, now it's sent
once. (The effects were not visible in the UI because both writes had
the same `span_id`.)
- Number of EFD retries is based on the time the _first_ attempt takes,
not the total time of the test.
- Faulty session logic should consider the number of new tests, not
known tests.
- Tests skipped by ITR should never be retried.

## Testing

Unit tests.

## Risks

None.

## Additional Notes

Since we now set tags directly from the retry handler, we might as well
do this for every tag instead of returning them from `get_final_status`
and `get_tags_for_test_run`. This is left as an exercise to the reader.