perf(debugger): reduce code origin startup time (#15272)
## Description

We adopt a lazy wrapping approach in Code Origin to reduce the startup
cost when instrumenting view functions. With the lazy approach, the
heavy instrumentation is performed when the view function is invoked for
the first time, preventing a potentially large delay on boot.

refs: DEBUG-4605

## Testing

Added benchmark to validate the performance improvements

## Risks

N/A

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->

---------

Co-authored-by: Gabriele N. Tornetta <gabriele.tornetta@datadoghq.com>
Co-authored-by: Gabriele N. Tornetta <P403n1x87@users.noreply.github.com>