fix(writer): avoid flushing while forking [backport 3.19] (#15661)
Fix a race condition where `_exporter.send` between the execution of
`stop_workers` in the before fork hook and the actual fork takes place.
Since `send` recreates the tokio runtime if it has been stopped, this
undoes the `stop_workers` call and causes the writer to deadlock in
child.

Also fixes a potential deadlock happening when attempting to lock the
`_request` Event in periodic service `stop` after a fork.

This has been tested locally in an airflow cluster which has been shown
to deadlock.

Co-authored-by: Brett Langdon <brett.langdon@datadoghq.com>