chore(ci_visibility): report flaky tests in new plugin (#15651)
## Description

This PR:

- Changes the retry handler API to set tags directly on the tests,
instead of returning tag dictionaries, as suggested in #15641;
- Adds an "is flaky run" property to tests. If a test has multiple
outcomes during EFD, it gets marked as a flaky run. This is then used to
log the test as `FLAKY` like the old plugin (instead of `PASSED` like we
were doing up to now).
- Moves retry report logging to a separate class that collects the retry
attempt reports as they are logged. When generating the final report,
the longrepr of the final report is copied from the longrepr of the
first attempt with the same outcome (e.g., if the final report is
`failed`, it will have the longrepr of the first failing report).
Previously the longrepr of the initial attempt was used, regardless of
its outcome (so if the first attempt was 'skipped' and the final report
was 'failed', this could lead to a failed report with a skipped
longrepr).
- If a test has failed attempt reports but the final outcome is not
failed, we still collect one failed report to present at the end of the
test session, so that at least one exception will be logged for the
test.

## Testing

Unit tests.

## Risks

None.

## Additional Notes

None.