fix(debugging): isDefined implementation (#15643)
## Description

We fix the implementation of the isDefined operation to accept an
arbitrary expression that is evaluated within the scope of a proxy
function. This is used to validate whether a certain expression would
evaluate to a value without errors.

(cherry picked from commit 4d1b76c4ea24cffa195a2447693e92371d83e215)